Moved `ImageGallery` to recipe-detail folder, and `SortControl` to recipe folder, and updated imports accordingly.
