+++ app/components/recipe-detail/ImageGallery.js
